Login app/requirements.txt
Pillow
streamlit
matplotlib
numpy
scipy
fpdf
pandas
